requirements.txt
torch>=1.8.0
torch-scatter>=2.0.8
torch-sparse>=0.6.12
torch-geometric>=2.0.0
ninja
easydict
pyyaml
dgl

